Iteratively do sufficiency tests on k features out of the given features

This function iteratively generates all combinations of k features from the given feature vector (n choose k), and perform a sufficiency test for each combination of the k features.

Usage

iterQCA(
  features,
  k,
  dataCali,
  outcome,
  incl.cut = 0.8,
  pri.cut = 0.51,
  n.cut = 10,
  sort.by = c("OUT", "incl"),
  complete = T
)

Arguments

features

string vector, a vector of features to choose from

k

number, number of features to choose for each sufficiency test

dataCali

dataframe, calibrated data

outcome

string, name of the outcome variable

incl.cut

inclusion cut-off(s); see `?QCA::truthTable` for details

pri.cut

minimal score for the PRI; see `?QCA::truthTable` for details

n.cut

minimum number of cases for a remainder row; see `?QCA::truthTable` for details

sort.by

Sort the truth table according to various columns

complete

Logical, whether to print complete truth table

Value

a summary table of solutions to all combinations of k out of n feautres

Examples

if (FALSE) { # \dontrun{
fDf <- getFeatureImp(xgb, voteData, "vote")
fVec <- head(fDf$feature, 10)
iterQCA(features = fVec, k=4, dataCali = voteDataCalibrated, outcome = "vote")
} # }