Perform sufficiency test on a select set of variables

Perform sufficiency test on a select set of variables

Usage

subsetSuffiTest(
  dataCali,
  vSelect,
  outcome,
  incl.cut = 0.8,
  pri.cut = 0.51,
  n.cut = 10,
  sort.by = c("OUT", "incl"),
  complete = T
)

Arguments

dataCali

dataframe, calibrated data

vSelect

string vector, selected variable names (must be in dataCali)

outcome

string, name of the outcome variable

incl.cut

inclusion cut-off(s); see `?QCA::truthTable` for details

pri.cut

minimal score for the PRI; see `?QCA::truthTable` for details

n.cut

minimum number of cases for a remainder row; see `?QCA::truthTable` for details

sort.by

Sort the truth table according to various columns

complete

Logical, whether to print complete truth table

Value

a list of two tables

table: truth table

path: sufficiency test solution

Examples

if (FALSE) { # \dontrun{
subsetSuffiTest(voteDataCalibrated, c("P_V530", "V648", "V207", "V141"), "vote")
} # }