helper function to combine results from a series of sufficiency tests

helper function to combine results from a series of sufficiency tests

Usage

sumSuffTestRes(featureList, resList)

Arguments

featureList

a list of feature vectors

resList

a list of corresponding results from `subsetSuffiTest()` on the above featureList

Value

dataframe, that summarizes the results from the sufficiency tests